Modified the design of the dashboard page to be much more responsive Improved PerformanceMetrics component responsive design Deleted the dashboard page group route Modified layout.js styling to accomodate the responsiveness of the child pages
